msm_fb: HDMI: Enable SPD InfoFrame transmission
Sinks use the information in the SPD InfoFrame to get
the vendor name and the product information of the source
device. This change adds the ability to specify this
information and transmit the data as part of the SPD
InfoFrame.

CRs-fixed: 388272
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>
(cherry picked from commit fff7a7d207a8791a68706bf8fac3838884981204)

Change-Id: I898473018d446e70cb821a5fc65ce55efb11f379
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>